chore: refactor pixel tests to seperate directory, remove reduntant test data
