+++ test/integrations/sources/shopify/data.ts
@@ -1,4599 +1,8 @@
 import { skip } from 'node:test';
-import utils from '../../../../src/v0/util';
-const defaultMockFns = () => {
-  jest.spyOn(utils, 'generateUUID').mockReturnValue('97fcd7b2-cc24-47d7-b776-057b7b199513');
-};
-
-const dummySourceConfig = {
-  ID: 'dummy-source-id',
-  OriginalID: '',
-  Name: 'Shopify v2',
-  Config: {
-    disableClientSideIdentifier: false,
-    eventUpload: false,
-    version: 'pixel',
-  },
-  Enabled: true,
-  WorkspaceID: 'dummy-workspace-id',
-  Destinations: null,
-  WriteKey: 'dummy-write-key',
-};
-
-const pixelEventsTestScenarios = [
-  {
-    name: 'shopify',
-    description: 'Page Call -> page_view event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f6b6f548-5FEF-4DAE-9CAB-39EE6F94E09B',
-              name: 'page_viewed',
-              data: {},
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T17:24:30.373Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                  characterSet: 'UTF-8',
-                  title: 'pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['page_viewed'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer:
-                        'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                      characterSet: 'UTF-8',
-                      title: 'pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'pixel-testing-rs',
-                      url: 'https://store.myshopify.com/',
-                      path: '/',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'page_viewed',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  name: 'Page View',
-                  type: 'page',
-                  properties: {},
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> product_viewed event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f6c07b5a-D20A-4E5F-812E-337299B56C34',
-              name: 'product_viewed',
-              data: {
-                productVariant: {
-                  price: {
-                    amount: 749.95,
-                    currencyCode: 'USD',
-                  },
-                  product: {
-                    title: 'The Collection Snowboard: Liquid',
-                    vendor: 'Hydrogen Vendor',
-                    id: '7234590834801',
-                    untranslatedTitle: 'The Collection Snowboard: Liquid',
-                    url: '/products/the-collection-snowboard-liquid',
-                    type: 'snowboard',
-                  },
-                  id: '41327143321713',
-                  image: {
-                    src: '//store.myshopify.com/cdn/shop/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                  },
-                  sku: '',
-                  title: 'Default Title',
-                  untranslatedTitle: 'Default Title',
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T17:34:54.889Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/products/the-collection-snowboard-liquid',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/',
-                  characterSet: 'UTF-8',
-                  title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/products/the-collection-snowboard-liquid',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['product_viewed'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/products/the-collection-snowboard-liquid',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/',
-                      characterSet: 'UTF-8',
-                      title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/products/the-collection-snowboard-liquid',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                      path: '/products/the-collection-snowboard-liquid',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'product_viewed',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Product Viewed',
-                  properties: {
-                    productVariant: {
-                      price: {
-                        amount: 749.95,
-                        currencyCode: 'USD',
-                      },
-                      product: {
-                        title: 'The Collection Snowboard: Liquid',
-                        vendor: 'Hydrogen Vendor',
-                        id: '7234590834801',
-                        untranslatedTitle: 'The Collection Snowboard: Liquid',
-                        url: '/products/the-collection-snowboard-liquid',
-                        type: 'snowboard',
-                      },
-                      id: '41327143321713',
-                      image: {
-                        src: '//store.myshopify.com/cdn/shop/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                      },
-                      sku: '',
-                      title: 'Default Title',
-                      untranslatedTitle: 'Default Title',
-                    },
-                    product_id: '7234590834801',
-                    variant: 'The Collection Snowboard: Liquid',
-                    brand: 'Hydrogen Vendor',
-                    category: 'snowboard',
-                    price: 749.95,
-                    currency: 'USD',
-                    url: '/products/the-collection-snowboard-liquid',
-                    name: 'The Collection Snowboard: Liquid',
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> cart_viewed event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'shu-f6eecef1-4132-459F-CDB5-681DA3DD61CD',
-              name: 'cart_viewed',
-              data: {
-                cart: {
-                  cost: {
-                    totalAmount: {
-                      amount: 1259.9,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  lines: [
-                    {
-                      cost: {
-                        totalAmount: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                      },
-                      merchandise: {
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          id: '7234590736497',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                          type: 'snowboard',
-                        },
-                        id: '41327143157873',
-                        image: {
-                          src: '//store.myshopify.com/cdn/shop/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a.jpg?v=1724736597',
-                        },
-                        sku: 'sku-managed-1',
-                        title: 'Default Title',
-                        untranslatedTitle: 'Default Title',
-                      },
-                      quantity: 2,
-                    },
-                  ],
-                  totalQuantity: 2,
-                  attributes: [],
-                  id: 'Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T18:25:30.125Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/cart',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/cart',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: '',
-                  characterSet: 'UTF-8',
-                  title: 'Your Shopping Cart – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/cart',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/cart',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['cart_viewed'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/cart',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/cart',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: '',
-                      characterSet: 'UTF-8',
-                      title: 'Your Shopping Cart – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/cart',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/cart',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Your Shopping Cart – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/cart',
-                      path: '/cart',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'cart_viewed',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Cart Viewed',
-                  properties: {
-                    products: [
-                      {
-                        cost: {
-                          totalAmount: {
-                            amount: 1259.9,
-                            currencyCode: 'USD',
-                          },
-                        },
-                        merchandise: {
-                          price: {
-                            amount: 629.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            title: 'The Multi-managed Snowboard',
-                            vendor: 'Multi-managed Vendor',
-                            id: '7234590736497',
-                            untranslatedTitle: 'The Multi-managed Snowboard',
-                            url: '/products/the-multi-managed-snowboard',
-                            type: 'snowboard',
-                          },
-                          id: '41327143157873',
-                          image: {
-                            src: '//store.myshopify.com/cdn/shop/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a.jpg?v=1724736597',
-                          },
-                          sku: 'sku-managed-1',
-                          title: 'Default Title',
-                          untranslatedTitle: 'Default Title',
-                        },
-                        quantity: 2,
-                        product_id: '7234590736497',
-                        variant: 'The Multi-managed Snowboard',
-                        image_url:
-                          '//store.myshopify.com/cdn/shop/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a.jpg?v=1724736597',
-                        price: 629.95,
-                        category: 'snowboard',
-                        url: '/products/the-multi-managed-snowboard',
-                        brand: 'Multi-managed Vendor',
-                        sku: 'sku-managed-1',
-                        name: 'Default Title',
-                      },
-                    ],
-                    cart_id: 'Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    total: 1259.9,
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> collection_viewed event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f6f0c6be-43F8-47D2-5F94-C22AD5ED3E79',
-              name: 'collection_viewed',
-              data: {
-                collection: {
-                  id: '',
-                  title: 'Products',
-                  productVariants: [
-                    {
-                      price: {
-                        amount: 10,
-                        currencyCode: 'USD',
-                      },
-                      product: {
-                        title: 'Gift Card',
-                        vendor: 'Snowboard Vendor',
-                        id: '7234590605425',
-                        untranslatedTitle: 'Gift Card',
-                        url: '/products/gift-card',
-                        type: 'giftcard',
-                      },
-                      id: '41327142895729',
-                      image: {
-                        src: '//store.myshopify.com/cdn/shop/files/gift_card.png?v=1724736596',
-                      },
-                      sku: '',
-                      title: '$10',
-                      untranslatedTitle: '$10',
-                    },
-                    {
-                      price: {
-                        amount: 24.95,
-                        currencyCode: 'USD',
-                      },
-                      product: {
-                        title: 'Selling Plans Ski Wax',
-                        vendor: 'pixel-testing-rs',
-                        id: '7234590802033',
-                        untranslatedTitle: 'Selling Plans Ski Wax',
-                        url: '/products/selling-plans-ski-wax',
-                        type: 'accessories',
-                      },
-                      id: '41327143223409',
-                      image: {
-                        src: '//store.myshopify.com/cdn/shop/files/snowboard_wax.png?v=1724736599',
-                      },
-                      sku: '',
-                      title: 'Selling Plans Ski Wax',
-                      untranslatedTitle: 'Selling Plans Ski Wax',
-                    },
-                    {
-                      price: {
-                        amount: 2629.95,
-                        currencyCode: 'USD',
-                      },
-                      product: {
-                        title: 'The 3p Fulfilled Snowboard',
-                        vendor: 'pixel-testing-rs',
-                        id: '7234590703729',
-                        untranslatedTitle: 'The 3p Fulfilled Snowboard',
-                        url: '/products/the-3p-fulfilled-snowboard',
-                        type: 'snowboard',
-                      },
-                      id: '41327143125105',
-                      image: {
-                        src: '//store.myshopify.com/cdn/shop/files/Main_b9e0da7f-db89-4d41-83f0-7f417b02831d.jpg?v=1724736597',
-                      },
-                      sku: 'sku-hosted-1',
-                      title: 'Default Title',
-                      untranslatedTitle: 'Default Title',
-                    },
-                  ],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T18:27:39.197Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/collections/all',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/collections/all',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Products – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/collections/all',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/collections/all',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['collection_viewed'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/collections/all',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/collections/all',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Products – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/collections/all',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/collections/all',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Products – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/collections/all',
-                      path: '/collections/all',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'collection_viewed',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Product List Viewed',
-                  properties: {
-                    cart_id: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                    list_id: 'sh-f6f0c6be-43F8-47D2-5F94-C22AD5ED3E79',
-                    products: [
-                      {
-                        price: 10,
-                        product: {
-                          title: 'Gift Card',
-                          vendor: 'Snowboard Vendor',
-                          id: '7234590605425',
-                          untranslatedTitle: 'Gift Card',
-                          url: '/products/gift-card',
-                          type: 'giftcard',
-                        },
-                        id: '41327142895729',
-                        image: {
-                          src: '//store.myshopify.com/cdn/shop/files/gift_card.png?v=1724736596',
-                        },
-                        sku: '',
-                        title: '$10',
-                        untranslatedTitle: '$10',
-                        image_url:
-                          '//store.myshopify.com/cdn/shop/files/gift_card.png?v=1724736596',
-                        product_id: '7234590605425',
-                        variant: 'Gift Card',
-                        category: 'giftcard',
-                        url: '/products/gift-card',
-                        brand: 'Snowboard Vendor',
-                        name: '$10',
-                      },
-                      {
-                        price: 24.95,
-                        product: {
-                          title: 'Selling Plans Ski Wax',
-                          vendor: 'pixel-testing-rs',
-                          id: '7234590802033',
-                          untranslatedTitle: 'Selling Plans Ski Wax',
-                          url: '/products/selling-plans-ski-wax',
-                          type: 'accessories',
-                        },
-                        id: '41327143223409',
-                        image: {
-                          src: '//store.myshopify.com/cdn/shop/files/snowboard_wax.png?v=1724736599',
-                        },
-                        sku: '',
-                        title: 'Selling Plans Ski Wax',
-                        untranslatedTitle: 'Selling Plans Ski Wax',
-                        image_url:
-                          '//store.myshopify.com/cdn/shop/files/snowboard_wax.png?v=1724736599',
-                        product_id: '7234590802033',
-                        variant: 'Selling Plans Ski Wax',
-                        category: 'accessories',
-                        url: '/products/selling-plans-ski-wax',
-                        brand: 'pixel-testing-rs',
-                        name: 'Selling Plans Ski Wax',
-                      },
-                      {
-                        price: 2629.95,
-                        product: {
-                          title: 'The 3p Fulfilled Snowboard',
-                          vendor: 'pixel-testing-rs',
-                          id: '7234590703729',
-                          untranslatedTitle: 'The 3p Fulfilled Snowboard',
-                          url: '/products/the-3p-fulfilled-snowboard',
-                          type: 'snowboard',
-                        },
-                        id: '41327143125105',
-                        image: {
-                          src: '//store.myshopify.com/cdn/shop/files/Main_b9e0da7f-db89-4d41-83f0-7f417b02831d.jpg?v=1724736597',
-                        },
-                        sku: 'sku-hosted-1',
-                        title: 'Default Title',
-                        untranslatedTitle: 'Default Title',
-                        image_url:
-                          '//store.myshopify.com/cdn/shop/files/Main_b9e0da7f-db89-4d41-83f0-7f417b02831d.jpg?v=1724736597',
-                        product_id: '7234590703729',
-                        variant: 'The 3p Fulfilled Snowboard',
-                        category: 'snowboard',
-                        url: '/products/the-3p-fulfilled-snowboard',
-                        brand: 'pixel-testing-rs',
-                        name: 'Default Title',
-                      },
-                    ],
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> product_added_to_cart event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f6f828db-F77B-43E8-96C4-1D51DACD52A3',
-              name: 'product_added_to_cart',
-              data: {
-                cartLine: {
-                  cost: {
-                    totalAmount: {
-                      amount: 749.95,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  merchandise: {
-                    id: '41327143321713',
-                    image: {
-                      src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                    },
-                    price: {
-                      amount: 749.95,
-                      currencyCode: 'USD',
-                    },
-                    product: {
-                      id: '7234590834801',
-                      title: 'The Collection Snowboard: Liquid',
-                      vendor: 'Hydrogen Vendor',
-                      type: 'snowboard',
-                      untranslatedTitle: 'The Collection Snowboard: Liquid',
-                      url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                    },
-                    sku: '',
-                    title: null,
-                    untranslatedTitle: null,
-                  },
-                  quantity: 1,
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T18:34:42.625Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/products/the-collection-snowboard-liquid',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/collections/all',
-                  characterSet: 'UTF-8',
-                  title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/products/the-collection-snowboard-liquid',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['carts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/products/the-collection-snowboard-liquid',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/collections/all',
-                      characterSet: 'UTF-8',
-                      title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/products/the-collection-snowboard-liquid',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'The Collection Snowboard: Liquid – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/products/the-collection-snowboard-liquid',
-                      path: '/products/the-collection-snowboard-liquid',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'product_added_to_cart',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Product Added',
-                  properties: {
-                    cartLine: {
-                      cost: {
-                        totalAmount: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                      },
-                      merchandise: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                        },
-                        sku: '',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      quantity: 1,
-                    },
-                    image_url:
-                      'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                    price: 749.95,
-                    product_id: '7234590834801',
-                    variant: 'The Collection Snowboard: Liquid',
-                    category: 'snowboard',
-                    brand: 'Hydrogen Vendor',
-                    url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                    sku: '',
-                    name: null,
-                    quantity: 1,
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> product_removed_from_cart event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'shu-f778d1eb-9B83-4832-9DC0-5C3B33A809F0',
-              name: 'product_removed_from_cart',
-              data: {
-                cartLine: {
-                  cost: {
-                    totalAmount: {
-                      amount: 749.95,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  merchandise: {
-                    id: '41327143321713',
-                    image: {
-                      src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                    },
-                    price: {
-                      amount: 749.95,
-                      currencyCode: 'USD',
-                    },
-                    product: {
-                      id: '7234590834801',
-                      title: 'The Collection Snowboard: Liquid',
-                      vendor: 'Hydrogen Vendor',
-                      type: 'snowboard',
-                      untranslatedTitle: 'The Collection Snowboard: Liquid',
-                      url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                    },
-                    sku: '',
-                    title: null,
-                    untranslatedTitle: null,
-                  },
-                  quantity: 1,
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T20:56:00.125Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/cart',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/cart',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: '',
-                  characterSet: 'UTF-8',
-                  title: 'Your Shopping Cart – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/cart',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/cart',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['carts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/cart',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/cart',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: '',
-                      characterSet: 'UTF-8',
-                      title: 'Your Shopping Cart – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/cart',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/cart',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Your Shopping Cart – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/cart',
-                      path: '/cart',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'product_removed_from_cart',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Product Removed',
-                  properties: {
-                    cartLine: {
-                      cost: {
-                        totalAmount: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                      },
-                      merchandise: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                        },
-                        sku: '',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      quantity: 1,
-                    },
-                    image_url:
-                      'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6.jpg?v=1724736600',
-                    price: 749.95,
-                    product_id: '7234590834801',
-                    variant: 'The Collection Snowboard: Liquid',
-                    category: 'snowboard',
-                    brand: 'Hydrogen Vendor',
-                    url: '/products/the-collection-snowboard-liquid?variant=41327143321713',
-                    sku: '',
-                    name: null,
-                    quantity: 1,
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> checkout_started event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
-              name: 'checkout_started',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: '',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [],
-                  },
-                  shippingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T20:57:59.674Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'checkout_started',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Started',
-                  properties: {
-                    products: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: 'The Collection Snowboard: Liquid',
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                        name: 'The Collection Snowboard: Liquid',
-                        image_url:
-                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        price: 749.95,
-                        sku: null,
-                        product_id: '7234590834801',
-                        category: 'snowboard',
-                        url: '/products/the-collection-snowboard-liquid',
-                        brand: 'Hydrogen Vendor',
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: 'The Multi-managed Snowboard',
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                        name: 'The Multi-managed Snowboard',
-                        image_url:
-                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        price: 629.95,
-                        sku: 'sku-managed-1',
-                        product_id: '7234590736497',
-                        category: 'snowboard',
-                        url: '/products/the-multi-managed-snowboard',
-                        brand: 'Multi-managed Vendor',
-                      },
-                    ],
-                    order_id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
-                    checkout_id: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    total: 2759.8,
-                    currency: 'USD',
-                    discount: 0,
-                    shipping: 0,
-                    revenue: 2759.8,
-                    value: 2759.8,
-                    tax: 0,
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> address_info_submitted event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7d2154d-7525-47A4-87FA-E54D2322E129',
-              name: 'checkout_address_info_submitted',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: 'test-user@sampleemail.com',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [
-                      {
-                        cost: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        costAfterDiscounts: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        description: null,
-                        handle: '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                        title: 'Economy',
-                        type: 'shipping',
-                      },
-                    ],
-                  },
-                  shippingAddress: {
-                    address1: 'Queens Center',
-                    address2: null,
-                    city: 'Elmhurst',
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: 'test',
-                    lastName: 'user',
-                    phone: null,
-                    province: 'NY',
-                    provinceCode: 'NY',
-                    zip: '11373',
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T21:45:50.523Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'checkout_address_info_submitted',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Address Info Submitted',
-                  properties: {
-                    buyerAcceptsEmailMarketing: false,
-                    buyerAcceptsSmsMarketing: false,
-                    attributes: [],
-                    billingAddress: {
-                      address1: null,
-                      address2: null,
-                      city: null,
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: null,
-                      lastName: null,
-                      phone: null,
-                      province: null,
-                      provinceCode: null,
-                      zip: null,
-                    },
-                    token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    currencyCode: 'USD',
-                    discountApplications: [],
-                    discountsAmount: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    email: 'test-user@sampleemail.com',
-                    phone: '',
-                    lineItems: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: {
-                          id: '41327143321713',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                          },
-                          price: {
-                            amount: 749.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590834801',
-                            title: 'The Collection Snowboard: Liquid',
-                            vendor: 'Hydrogen Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Collection Snowboard: Liquid',
-                            url: '/products/the-collection-snowboard-liquid',
-                          },
-                          sku: null,
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: {
-                          id: '41327143157873',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                          },
-                          price: {
-                            amount: 629.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590736497',
-                            title: 'The Multi-managed Snowboard',
-                            vendor: 'Multi-managed Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Multi-managed Snowboard',
-                            url: '/products/the-multi-managed-snowboard',
-                          },
-                          sku: 'sku-managed-1',
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                    ],
-                    localization: {
-                      country: {
-                        isoCode: 'US',
-                      },
-                      language: {
-                        isoCode: 'en-US',
-                      },
-                      market: {
-                        id: 'gid://shopify/Market/23505895537',
-                        handle: 'us',
-                      },
-                    },
-                    order: {
-                      id: null,
-                      customer: {
-                        id: null,
-                        isFirstOrder: null,
-                      },
-                    },
-                    delivery: {
-                      selectedDeliveryOptions: [
-                        {
-                          cost: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          costAfterDiscounts: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          description: null,
-                          handle:
-                            '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                          title: 'Economy',
-                          type: 'shipping',
-                        },
-                      ],
-                    },
-                    shippingAddress: {
-                      address1: 'Queens Center',
-                      address2: null,
-                      city: 'Elmhurst',
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: 'test',
-                      lastName: 'user',
-                      phone: null,
-                      province: 'NY',
-                      provinceCode: 'NY',
-                      zip: '11373',
-                    },
-                    subtotalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    shippingLine: {
-                      price: {
-                        amount: 0,
-                        currencyCode: 'USD',
-                      },
-                    },
-                    smsMarketingPhone: null,
-                    totalTax: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    totalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    transactions: [],
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> contact_info_submitted event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7c8416f-1D35-4304-EF29-78666678C4E9',
-              name: 'checkout_contact_info_submitted',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: 'test-user@sampleemail.com',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [],
-                  },
-                  shippingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T21:40:28.498Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'checkout_contact_info_submitted',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Contact Info Submitted',
-                  properties: {
-                    buyerAcceptsEmailMarketing: false,
-                    buyerAcceptsSmsMarketing: false,
-                    attributes: [],
-                    billingAddress: {
-                      address1: null,
-                      address2: null,
-                      city: null,
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: null,
-                      lastName: null,
-                      phone: null,
-                      province: null,
-                      provinceCode: null,
-                      zip: null,
-                    },
-                    token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    currencyCode: 'USD',
-                    discountApplications: [],
-                    discountsAmount: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    email: 'test-user@sampleemail.com',
-                    phone: '',
-                    lineItems: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: {
-                          id: '41327143321713',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                          },
-                          price: {
-                            amount: 749.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590834801',
-                            title: 'The Collection Snowboard: Liquid',
-                            vendor: 'Hydrogen Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Collection Snowboard: Liquid',
-                            url: '/products/the-collection-snowboard-liquid',
-                          },
-                          sku: null,
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: {
-                          id: '41327143157873',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                          },
-                          price: {
-                            amount: 629.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590736497',
-                            title: 'The Multi-managed Snowboard',
-                            vendor: 'Multi-managed Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Multi-managed Snowboard',
-                            url: '/products/the-multi-managed-snowboard',
-                          },
-                          sku: 'sku-managed-1',
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                    ],
-                    localization: {
-                      country: {
-                        isoCode: 'US',
-                      },
-                      language: {
-                        isoCode: 'en-US',
-                      },
-                      market: {
-                        id: 'gid://shopify/Market/23505895537',
-                        handle: 'us',
-                      },
-                    },
-                    order: {
-                      id: null,
-                      customer: {
-                        id: null,
-                        isFirstOrder: null,
-                      },
-                    },
-                    delivery: {
-                      selectedDeliveryOptions: [],
-                    },
-                    shippingAddress: {
-                      address1: null,
-                      address2: null,
-                      city: null,
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: null,
-                      lastName: null,
-                      phone: null,
-                      province: null,
-                      provinceCode: null,
-                      zip: null,
-                    },
-                    subtotalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    shippingLine: {
-                      price: {
-                        amount: 0,
-                        currencyCode: 'USD',
-                      },
-                    },
-                    smsMarketingPhone: null,
-                    totalTax: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    totalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    transactions: [],
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> shipping_info_submitted event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7d5618e-404A-4E6D-4662-599A4BCC9E7C',
-              name: 'checkout_shipping_info_submitted',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: 'test-user@sampleemail.com',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [
-                      {
-                        cost: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        costAfterDiscounts: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        description: null,
-                        handle: '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                        title: 'Economy',
-                        type: 'shipping',
-                      },
-                    ],
-                  },
-                  shippingAddress: {
-                    address1: 'Queens Center',
-                    address2: null,
-                    city: 'Elmhurst',
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: 'test',
-                    lastName: 'user',
-                    phone: null,
-                    province: 'NY',
-                    provinceCode: 'NY',
-                    zip: '11373',
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T21:47:38.576Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 433,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 433,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 433,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 433,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'checkout_shipping_info_submitted',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Shipping Info Submitted',
-                  properties: {
-                    buyerAcceptsEmailMarketing: false,
-                    buyerAcceptsSmsMarketing: false,
-                    attributes: [],
-                    billingAddress: {
-                      address1: null,
-                      address2: null,
-                      city: null,
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: null,
-                      lastName: null,
-                      phone: null,
-                      province: null,
-                      provinceCode: null,
-                      zip: null,
-                    },
-                    token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    currencyCode: 'USD',
-                    discountApplications: [],
-                    discountsAmount: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    email: 'test-user@sampleemail.com',
-                    phone: '',
-                    lineItems: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: {
-                          id: '41327143321713',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                          },
-                          price: {
-                            amount: 749.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590834801',
-                            title: 'The Collection Snowboard: Liquid',
-                            vendor: 'Hydrogen Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Collection Snowboard: Liquid',
-                            url: '/products/the-collection-snowboard-liquid',
-                          },
-                          sku: null,
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: {
-                          id: '41327143157873',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                          },
-                          price: {
-                            amount: 629.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590736497',
-                            title: 'The Multi-managed Snowboard',
-                            vendor: 'Multi-managed Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Multi-managed Snowboard',
-                            url: '/products/the-multi-managed-snowboard',
-                          },
-                          sku: 'sku-managed-1',
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                    ],
-                    localization: {
-                      country: {
-                        isoCode: 'US',
-                      },
-                      language: {
-                        isoCode: 'en-US',
-                      },
-                      market: {
-                        id: 'gid://shopify/Market/23505895537',
-                        handle: 'us',
-                      },
-                    },
-                    order: {
-                      id: null,
-                      customer: {
-                        id: null,
-                        isFirstOrder: null,
-                      },
-                    },
-                    delivery: {
-                      selectedDeliveryOptions: [
-                        {
-                          cost: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          costAfterDiscounts: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          description: null,
-                          handle:
-                            '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                          title: 'Economy',
-                          type: 'shipping',
-                        },
-                      ],
-                    },
-                    shippingAddress: {
-                      address1: 'Queens Center',
-                      address2: null,
-                      city: 'Elmhurst',
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: 'test',
-                      lastName: 'user',
-                      phone: null,
-                      province: 'NY',
-                      provinceCode: 'NY',
-                      zip: '11373',
-                    },
-                    subtotalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    shippingLine: {
-                      price: {
-                        amount: 0,
-                        currencyCode: 'USD',
-                      },
-                    },
-                    smsMarketingPhone: null,
-                    totalTax: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    totalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    transactions: [],
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> payment_info_submitted event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7d843ea-ED11-4A12-F32F-C5A45BED0413',
-              name: 'payment_info_submitted',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: 'test-user@sampleemail.com',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [
-                      {
-                        cost: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        costAfterDiscounts: {
-                          amount: 0,
-                          currencyCode: 'USD',
-                        },
-                        description: null,
-                        handle: '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                        title: 'Economy',
-                        type: 'shipping',
-                      },
-                    ],
-                  },
-                  shippingAddress: {
-                    address1: 'Queens Center',
-                    address2: null,
-                    city: 'Elmhurst',
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: 'test',
-                    lastName: 'user',
-                    phone: null,
-                    province: 'NY',
-                    provinceCode: 'NY',
-                    zip: '11373',
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T21:49:13.092Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname:
-                      '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname:
-                      '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU/processing',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'payment_info_submitted',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Payment Info Submitted',
-                  properties: {
-                    buyerAcceptsEmailMarketing: false,
-                    buyerAcceptsSmsMarketing: false,
-                    attributes: [],
-                    billingAddress: {
-                      address1: null,
-                      address2: null,
-                      city: null,
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: null,
-                      lastName: null,
-                      phone: null,
-                      province: null,
-                      provinceCode: null,
-                      zip: null,
-                    },
-                    token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    currencyCode: 'USD',
-                    discountApplications: [],
-                    discountsAmount: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    email: 'test-user@sampleemail.com',
-                    phone: '',
-                    lineItems: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: {
-                          id: '41327143321713',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                          },
-                          price: {
-                            amount: 749.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590834801',
-                            title: 'The Collection Snowboard: Liquid',
-                            vendor: 'Hydrogen Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Collection Snowboard: Liquid',
-                            url: '/products/the-collection-snowboard-liquid',
-                          },
-                          sku: null,
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: {
-                          id: '41327143157873',
-                          image: {
-                            src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                          },
-                          price: {
-                            amount: 629.95,
-                            currencyCode: 'USD',
-                          },
-                          product: {
-                            id: '7234590736497',
-                            title: 'The Multi-managed Snowboard',
-                            vendor: 'Multi-managed Vendor',
-                            type: 'snowboard',
-                            untranslatedTitle: 'The Multi-managed Snowboard',
-                            url: '/products/the-multi-managed-snowboard',
-                          },
-                          sku: 'sku-managed-1',
-                          title: null,
-                          untranslatedTitle: null,
-                        },
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                      },
-                    ],
-                    localization: {
-                      country: {
-                        isoCode: 'US',
-                      },
-                      language: {
-                        isoCode: 'en-US',
-                      },
-                      market: {
-                        id: 'gid://shopify/Market/23505895537',
-                        handle: 'us',
-                      },
-                    },
-                    order: {
-                      id: null,
-                      customer: {
-                        id: null,
-                        isFirstOrder: null,
-                      },
-                    },
-                    delivery: {
-                      selectedDeliveryOptions: [
-                        {
-                          cost: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          costAfterDiscounts: {
-                            amount: 0,
-                            currencyCode: 'USD',
-                          },
-                          description: null,
-                          handle:
-                            '5f7028e0bd5225c17b24bdaa0c09f914-8388085074acab7e91de633521be86f0',
-                          title: 'Economy',
-                          type: 'shipping',
-                        },
-                      ],
-                    },
-                    shippingAddress: {
-                      address1: 'Queens Center',
-                      address2: null,
-                      city: 'Elmhurst',
-                      country: 'US',
-                      countryCode: 'US',
-                      firstName: 'test',
-                      lastName: 'user',
-                      phone: null,
-                      province: 'NY',
-                      provinceCode: 'NY',
-                      zip: '11373',
-                    },
-                    subtotalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    shippingLine: {
-                      price: {
-                        amount: 0,
-                        currencyCode: 'USD',
-                      },
-                    },
-                    smsMarketingPhone: null,
-                    totalTax: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                    totalPrice: {
-                      amount: 2759.8,
-                      currencyCode: 'USD',
-                    },
-                    transactions: [],
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> checkout_completed event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
-              name: 'checkout_completed',
-              data: {
-                checkout: {
-                  buyerAcceptsEmailMarketing: false,
-                  buyerAcceptsSmsMarketing: false,
-                  attributes: [],
-                  billingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
-                  currencyCode: 'USD',
-                  discountApplications: [],
-                  discountsAmount: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  email: '',
-                  phone: '',
-                  lineItems: [
-                    {
-                      discountAllocations: [],
-                      id: '41327143321713',
-                      quantity: 2,
-                      title: 'The Collection Snowboard: Liquid',
-                      variant: {
-                        id: '41327143321713',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        },
-                        price: {
-                          amount: 749.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590834801',
-                          title: 'The Collection Snowboard: Liquid',
-                          vendor: 'Hydrogen Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Collection Snowboard: Liquid',
-                          url: '/products/the-collection-snowboard-liquid',
-                        },
-                        sku: null,
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1499.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                    {
-                      discountAllocations: [],
-                      id: '41327143157873',
-                      quantity: 2,
-                      title: 'The Multi-managed Snowboard',
-                      variant: {
-                        id: '41327143157873',
-                        image: {
-                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        },
-                        price: {
-                          amount: 629.95,
-                          currencyCode: 'USD',
-                        },
-                        product: {
-                          id: '7234590736497',
-                          title: 'The Multi-managed Snowboard',
-                          vendor: 'Multi-managed Vendor',
-                          type: 'snowboard',
-                          untranslatedTitle: 'The Multi-managed Snowboard',
-                          url: '/products/the-multi-managed-snowboard',
-                        },
-                        sku: 'sku-managed-1',
-                        title: null,
-                        untranslatedTitle: null,
-                      },
-                      finalLinePrice: {
-                        amount: 1259.9,
-                        currencyCode: 'USD',
-                      },
-                      sellingPlanAllocation: null,
-                      properties: [],
-                    },
-                  ],
-                  localization: {
-                    country: {
-                      isoCode: 'US',
-                    },
-                    language: {
-                      isoCode: 'en-US',
-                    },
-                    market: {
-                      id: 'gid://shopify/Market/23505895537',
-                      handle: 'us',
-                    },
-                  },
-                  order: {
-                    id: null,
-                    customer: {
-                      id: null,
-                      isFirstOrder: null,
-                    },
-                  },
-                  delivery: {
-                    selectedDeliveryOptions: [],
-                  },
-                  shippingAddress: {
-                    address1: null,
-                    address2: null,
-                    city: null,
-                    country: 'US',
-                    countryCode: 'US',
-                    firstName: null,
-                    lastName: null,
-                    phone: null,
-                    province: null,
-                    provinceCode: null,
-                    zip: null,
-                  },
-                  subtotalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  shippingLine: {
-                    price: {
-                      amount: 0,
-                      currencyCode: 'USD',
-                    },
-                  },
-                  smsMarketingPhone: null,
-                  totalTax: {
-                    amount: 0,
-                    currencyCode: 'USD',
-                  },
-                  totalPrice: {
-                    amount: 2759.8,
-                    currencyCode: 'USD',
-                  },
-                  transactions: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T20:57:59.674Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  referrer: 'https://store.myshopify.com/cart',
-                  characterSet: 'UTF-8',
-                  title: 'Checkout - pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                    port: '',
-                    protocol: 'https:',
-                    search: '',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      referrer: 'https://store.myshopify.com/cart',
-                      characterSet: 'UTF-8',
-                      title: 'Checkout - pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname:
-                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                        port: '',
-                        protocol: 'https:',
-                        search: '',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Checkout - pixel-testing-rs',
-                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
-                      search: '',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'checkout_completed',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Completed',
-                  properties: {
-                    products: [
-                      {
-                        discountAllocations: [],
-                        id: '41327143321713',
-                        quantity: 2,
-                        title: 'The Collection Snowboard: Liquid',
-                        variant: 'The Collection Snowboard: Liquid',
-                        finalLinePrice: {
-                          amount: 1499.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                        name: 'The Collection Snowboard: Liquid',
-                        image_url:
-                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
-                        price: 749.95,
-                        sku: null,
-                        product_id: '7234590834801',
-                        category: 'snowboard',
-                        url: '/products/the-collection-snowboard-liquid',
-                        brand: 'Hydrogen Vendor',
-                      },
-                      {
-                        discountAllocations: [],
-                        id: '41327143157873',
-                        quantity: 2,
-                        title: 'The Multi-managed Snowboard',
-                        variant: 'The Multi-managed Snowboard',
-                        finalLinePrice: {
-                          amount: 1259.9,
-                          currencyCode: 'USD',
-                        },
-                        sellingPlanAllocation: null,
-                        properties: [],
-                        name: 'The Multi-managed Snowboard',
-                        image_url:
-                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
-                        price: 629.95,
-                        sku: 'sku-managed-1',
-                        product_id: '7234590736497',
-                        category: 'snowboard',
-                        url: '/products/the-multi-managed-snowboard',
-                        brand: 'Multi-managed Vendor',
-                      },
-                    ],
-                    order_id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
-                    checkout_id: '5f7028e0bd5225c17b24bdaa0c09f914',
-                    total: 2759.8,
-                    currency: 'USD',
-                    discount: 0,
-                    shipping: 0,
-                    revenue: 2759.8,
-                    value: 2759.8,
-                    tax: 0,
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> search_submitted event from web pixel',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7d599b4-D80F-4D05-C4CE-B980D5444596',
-              name: 'search_submitted',
-              data: {
-                searchResult: {
-                  query: 'skate',
-                  productVariants: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T22:37:35.869Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/search',
-                    port: '',
-                    protocol: 'https:',
-                    search: '?q=skate&options%5Bprefix%5D=last',
-                  },
-                  referrer: 'https://store.myshopify.com/',
-                  characterSet: 'UTF-8',
-                  title: 'Search: 0 results found for "skate" – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/search',
-                    port: '',
-                    protocol: 'https:',
-                    search: '?q=skate&options%5Bprefix%5D=last',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['search_submitted'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    document: {
-                      location: {
-                        href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/search',
-                        port: '',
-                        protocol: 'https:',
-                        search: '?q=skate&options%5Bprefix%5D=last',
-                      },
-                      referrer: 'https://store.myshopify.com/',
-                      characterSet: 'UTF-8',
-                      title: 'Search: 0 results found for "skate" – pixel-testing-rs',
-                    },
-                    navigator: {
-                      language: 'en-US',
-                      cookieEnabled: true,
-                      languages: ['en-US', 'en'],
-                      userAgent:
-                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    },
-                    window: {
-                      innerHeight: 1028,
-                      innerWidth: 1362,
-                      outerHeight: 1080,
-                      outerWidth: 1728,
-                      pageXOffset: 0,
-                      pageYOffset: 0,
-                      location: {
-                        href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                        hash: '',
-                        host: 'store.myshopify.com',
-                        hostname: 'store.myshopify.com',
-                        origin: 'https://store.myshopify.com',
-                        pathname: '/search',
-                        port: '',
-                        protocol: 'https:',
-                        search: '?q=skate&options%5Bprefix%5D=last',
-                      },
-                      origin: 'https://store.myshopify.com',
-                      screen: {
-                        height: 1117,
-                        width: 1728,
-                      },
-                      screenX: 0,
-                      screenY: 37,
-                      scrollX: 0,
-                      scrollY: 0,
-                    },
-                    page: {
-                      title: 'Search: 0 results found for "skate" – pixel-testing-rs',
-                      url: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                      path: '/search',
-                      search: '?q=skate&options%5Bprefix%5D=last',
-                    },
-                    userAgent:
-                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                    screen: {
-                      height: 1117,
-                      width: 1728,
-                    },
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'client',
-                      version: '2.0.0',
-                    },
-                    topic: 'search_submitted',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Search Submitted',
-                  properties: {
-                    query: 'skate',
-                  },
-                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> unknown event from web pixel, should not be sent to Shopify',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'sh-f7d599b4-D80F-4D05-C4CE-B980D5444596',
-              name: 'unknown_event',
-              data: {
-                searchResult: {
-                  query: 'skate',
-                  productVariants: [],
-                },
-              },
-              type: 'standard',
-              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
-              timestamp: '2024-09-15T22:37:35.869Z',
-              context: {
-                document: {
-                  location: {
-                    href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/search',
-                    port: '',
-                    protocol: 'https:',
-                    search: '?q=skate&options%5Bprefix%5D=last',
-                  },
-                  referrer: 'https://store.myshopify.com/',
-                  characterSet: 'UTF-8',
-                  title: 'Search: 0 results found for "skate" – pixel-testing-rs',
-                },
-                navigator: {
-                  language: 'en-US',
-                  cookieEnabled: true,
-                  languages: ['en-US', 'en'],
-                  userAgent:
-                    'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
-                },
-                window: {
-                  innerHeight: 1028,
-                  innerWidth: 1362,
-                  outerHeight: 1080,
-                  outerWidth: 1728,
-                  pageXOffset: 0,
-                  pageYOffset: 0,
-                  location: {
-                    href: 'https://store.myshopify.com/search?q=skate&options%5Bprefix%5D=last',
-                    hash: '',
-                    host: 'store.myshopify.com',
-                    hostname: 'store.myshopify.com',
-                    origin: 'https://store.myshopify.com',
-                    pathname: '/search',
-                    port: '',
-                    protocol: 'https:',
-                    search: '?q=skate&options%5Bprefix%5D=last',
-                  },
-                  origin: 'https://store.myshopify.com',
-                  screen: {
-                    height: 1117,
-                    width: 1728,
-                  },
-                  screenX: 0,
-                  screenY: 37,
-                  scrollX: 0,
-                  scrollY: 0,
-                },
-              },
-              pixelEventLabel: true,
-              query_parameters: {
-                topic: ['search_submitted'],
-                writeKey: ['dummy-write-key'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            outputToSource: {
-              body: 'T0s=',
-              contentType: 'text/plain',
-            },
-            statusCode: 200,
-          },
-        ],
-      },
-    },
-  },
-];
+import { pixelCheckoutEventsTestScenarios } from './pixelTestScenarios/CheckoutEventsTests';
+import { pixelCheckoutStepsScenarios } from './pixelTestScenarios/CheckoutStepsTests';
+import { pixelEventsTestScenarios } from './pixelTestScenarios/ProductEventsTests';
+import { v1ServerSideEventsScenarios } from './v1ServerSideEventsTests';
 
 const serverSideEventsScenarios = [
   {
@@ -5203,598 +612,9 @@ const serverSideEventsScenarios = [
   },
 ];
 
-const v1ServerSideEventsScenarios = [
-  {
-    name: 'shopify',
-    description: 'Track Call -> Checkout Updated event',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 35374569160817,
-              token: 'e89d4437003b6b8480f8bc7f8036a659',
-              cart_token: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-              email: 'AJAIKUMARBAJPAI1962@GMAIL.COM',
-              gateway: null,
-              buyer_accepts_marketing: false,
-              buyer_accepts_sms_marketing: false,
-              sms_marketing_phone: null,
-              created_at: '2024-09-16T03:50:15+00:00',
-              updated_at: '2024-09-17T03:29:02-04:00',
-              landing_site: '/',
-              note: '',
-              note_attributes: [],
-              referring_site: '',
-              shipping_lines: [
-                {
-                  code: 'Standard',
-                  price: '6.90',
-                  original_shop_price: '6.90',
-                  original_shop_markup: '0.00',
-                  source: 'shopify',
-                  title: 'Standard',
-                  presentment_title: 'Standard',
-                  phone: null,
-                  tax_lines: [],
-                  custom_tax_lines: null,
-                  markup: '0.00',
-                  carrier_identifier: null,
-                  carrier_service_id: null,
-                  api_client_id: '580111',
-                  delivery_option_group: {
-                    token: '26492692a443ee35c30eb82073bacaa8',
-                    type: 'one_time_purchase',
-                  },
-                  delivery_expectation_range: null,
-                  delivery_expectation_type: null,
-                  id: null,
-                  requested_fulfillment_service_id: null,
-                  delivery_category: null,
-                  validation_context: null,
-                  applied_discounts: [],
-                },
-              ],
-              shipping_address: {
-                first_name: 'Yashasvi',
-                address1: '103 Block B',
-                phone: null,
-                city: 'Kanpur',
-                zip: '85003',
-                province: 'Arizona',
-                country: 'United States',
-                last_name: 'Bajpai',
-                address2: 'Shyam Nagar',
-                company: null,
-                latitude: null,
-                longitude: null,
-                name: 'Yashasvi Bajpai',
-                country_code: 'US',
-                province_code: 'AZ',
-              },
-              taxes_included: false,
-              total_weight: 0,
-              currency: 'USD',
-              completed_at: null,
-              phone: null,
-              customer_locale: 'en-US',
-              line_items: [
-                {
-                  key: '41327143059569',
-                  fulfillment_service: 'manual',
-                  gift_card: false,
-                  grams: 0,
-                  presentment_title: 'The Multi-location Snowboard',
-                  presentment_variant_title: '',
-                  product_id: 7234590638193,
-                  quantity: 1,
-                  requires_shipping: true,
-                  sku: '',
-                  tax_lines: [],
-                  taxable: true,
-                  title: 'The Multi-location Snowboard',
-                  variant_id: 41327143059569,
-                  variant_title: '',
-                  variant_price: '729.95',
-                  vendor: 'pixel-testing-rs',
-                  unit_price_measurement: {
-                    measured_type: null,
-                    quantity_value: null,
-                    quantity_unit: null,
-                    reference_value: null,
-                    reference_unit: null,
-                  },
-                  compare_at_price: null,
-                  line_price: '729.95',
-                  price: '729.95',
-                  applied_discounts: [],
-                  destination_location_id: null,
-                  user_id: null,
-                  rank: null,
-                  origin_location_id: null,
-                  properties: {},
-                },
-              ],
-              name: '#35374569160817',
-              abandoned_checkout_url:
-                'https://pixel-testing-rs.myshopify.com/59026964593/checkouts/ac/Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2/recover?key=8195f56ee0de230b3a0469cc692f3436',
-              discount_codes: [],
-              tax_lines: [],
-              presentment_currency: 'USD',
-              source_name: 'web',
-              total_line_items_price: '729.95',
-              total_tax: '0.00',
-              total_discounts: '0.00',
-              subtotal_price: '729.95',
-              total_price: '736.85',
-              total_duties: '0.00',
-              device_id: null,
-              user_id: null,
-              location_id: null,
-              source_identifier: null,
-              source_url: null,
-              source: null,
-              closed_at: null,
-              customer: {
-                id: 7188389789809,
-                email: 'ajaikumarbajpai1962@gmail.com',
-                accepts_marketing: false,
-                created_at: null,
-                updated_at: null,
-                first_name: 'Yashasvi',
-                last_name: 'Bajpai',
-                orders_count: 0,
-                state: 'disabled',
-                total_spent: '0.00',
-                last_order_id: null,
-                note: null,
-                verified_email: true,
-                multipass_identifier: null,
-                tax_exempt: false,
-                phone: null,
-                tags: '',
-                currency: 'USD',
-                accepts_marketing_updated_at: null,
-                admin_graphql_api_id: 'gid://shopify/Customer/7188389789809',
-                default_address: {
-                  id: null,
-                  customer_id: 7188389789809,
-                  first_name: 'Yashasvi',
-                  last_name: 'Bajpai',
-                  company: null,
-                  address1: '103 Block B',
-                  address2: 'Shyam Nagar',
-                  city: 'Kanpur',
-                  province: 'Arizona',
-                  country: 'United States',
-                  zip: '85003',
-                  phone: null,
-                  name: 'Yashasvi Bajpai',
-                  province_code: 'AZ',
-                  country_code: 'US',
-                  country_name: 'United States',
-                  default: true,
-                },
-                last_order_name: null,
-                marketing_opt_in_level: null,
-              },
-              query_parameters: {
-                topic: ['checkouts_update'],
-                writeKey: ['2l9QoM7KRMJLMcYhXNUVDT0Mqbd'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'server',
-                      version: '2.0.0',
-                    },
-                    integration: {
-                      name: 'SHOPIFY',
-                    },
-                    topic: 'checkouts_update',
-                    cart_token: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Checkout Updated',
-                  properties: {
-                    order_id: 35374569160817,
-                    value: '736.85',
-                    tax: '0.00',
-                    currency: 'USD',
-                    token: 'e89d4437003b6b8480f8bc7f8036a659',
-                    cart_token: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-                    email: 'AJAIKUMARBAJPAI1962@GMAIL.COM',
-                    buyer_accepts_marketing: false,
-                    buyer_accepts_sms_marketing: false,
-                    created_at: '2024-09-16T03:50:15+00:00',
-                    updated_at: '2024-09-17T03:29:02-04:00',
-                    landing_site: '/',
-                    note: '',
-                    note_attributes: [],
-                    referring_site: '',
-                    shipping_lines: [
-                      {
-                        code: 'Standard',
-                        price: '6.90',
-                        original_shop_price: '6.90',
-                        original_shop_markup: '0.00',
-                        source: 'shopify',
-                        title: 'Standard',
-                        presentment_title: 'Standard',
-                        phone: null,
-                        tax_lines: [],
-                        custom_tax_lines: null,
-                        markup: '0.00',
-                        carrier_identifier: null,
-                        carrier_service_id: null,
-                        api_client_id: '580111',
-                        delivery_option_group: {
-                          token: '26492692a443ee35c30eb82073bacaa8',
-                          type: 'one_time_purchase',
-                        },
-                        delivery_expectation_range: null,
-                        delivery_expectation_type: null,
-                        id: null,
-                        requested_fulfillment_service_id: null,
-                        delivery_category: null,
-                        validation_context: null,
-                        applied_discounts: [],
-                      },
-                    ],
-                    taxes_included: false,
-                    total_weight: 0,
-                    customer_locale: 'en-US',
-                    name: '#35374569160817',
-                    abandoned_checkout_url:
-                      'https://pixel-testing-rs.myshopify.com/59026964593/checkouts/ac/Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2/recover?key=8195f56ee0de230b3a0469cc692f3436',
-                    discount_codes: [],
-                    tax_lines: [],
-                    presentment_currency: 'USD',
-                    source_name: 'web',
-                    total_line_items_price: '729.95',
-                    total_discounts: '0.00',
-                    subtotal_price: '729.95',
-                    total_duties: '0.00',
-                    products: [
-                      {
-                        product_id: 7234590638193,
-                        price: '729.95',
-                        brand: 'pixel-testing-rs',
-                        quantity: 1,
-                        key: '41327143059569',
-                        fulfillment_service: 'manual',
-                        gift_card: false,
-                        grams: 0,
-                        presentment_title: 'The Multi-location Snowboard',
-                        presentment_variant_title: '',
-                        requires_shipping: true,
-                        tax_lines: [],
-                        taxable: true,
-                        title: 'The Multi-location Snowboard',
-                        unit_price_measurement: {
-                          measured_type: null,
-                          quantity_value: null,
-                          quantity_unit: null,
-                          reference_value: null,
-                          reference_unit: null,
-                        },
-                        compare_at_price: null,
-                        line_price: '729.95',
-                        applied_discounts: [],
-                        destination_location_id: null,
-                        user_id: null,
-                        rank: null,
-                        origin_location_id: null,
-                        properties: {},
-                        variant: '41327143059569 729.95 ',
-                      },
-                    ],
-                  },
-                  userId: '7188389789809',
-                  traits: {
-                    email: 'ajaikumarbajpai1962@gmail.com',
-                    firstName: 'Yashasvi',
-                    lastName: 'Bajpai',
-                    address: {
-                      id: null,
-                      customer_id: 7188389789809,
-                      first_name: 'Yashasvi',
-                      last_name: 'Bajpai',
-                      company: null,
-                      address1: '103 Block B',
-                      address2: 'Shyam Nagar',
-                      city: 'Kanpur',
-                      province: 'Arizona',
-                      country: 'United States',
-                      zip: '85003',
-                      phone: null,
-                      name: 'Yashasvi Bajpai',
-                      province_code: 'AZ',
-                      country_code: 'US',
-                      country_name: 'United States',
-                      default: true,
-                    },
-                    acceptsMarketing: false,
-                    orderCount: 0,
-                    state: 'disabled',
-                    totalSpent: '0.00',
-                    verifiedEmail: true,
-                    taxExempt: false,
-                    tags: '',
-                    currency: 'USD',
-                    adminGraphqlApiId: 'gid://shopify/Customer/7188389789809',
-                    shippingAddress: {
-                      first_name: 'Yashasvi',
-                      address1: '103 Block B',
-                      phone: null,
-                      city: 'Kanpur',
-                      zip: '85003',
-                      province: 'Arizona',
-                      country: 'United States',
-                      last_name: 'Bajpai',
-                      address2: 'Shyam Nagar',
-                      company: null,
-                      latitude: null,
-                      longitude: null,
-                      name: 'Yashasvi Bajpai',
-                      country_code: 'US',
-                      province_code: 'AZ',
-                    },
-                  },
-                  timestamp: '2024-09-17T07:29:02.000Z',
-                  anonymousId: '97fcd7b2-cc24-47d7-b776-057b7b199513',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-    mockFns: () => {
-      defaultMockFns();
-    },
-  },
-  {
-    name: 'shopify',
-    description: 'Track Call -> Cart Update event',
-    module: 'source',
-    version: 'v1',
-    input: {
-      request: {
-        body: [
-          {
-            event: {
-              id: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-              token: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-              line_items: [
-                {
-                  id: 41327143059569,
-                  properties: null,
-                  quantity: 3,
-                  variant_id: 41327143059569,
-                  key: '41327143059569:90562f18109e0e6484b0c297e7981b30',
-                  discounted_price: '729.95',
-                  discounts: [],
-                  gift_card: false,
-                  grams: 0,
-                  line_price: '2189.85',
-                  original_line_price: '2189.85',
-                  original_price: '729.95',
-                  price: '729.95',
-                  product_id: 7234590638193,
-                  sku: '',
-                  taxable: true,
-                  title: 'The Multi-location Snowboard',
-                  total_discount: '0.00',
-                  vendor: 'pixel-testing-rs',
-                  discounted_price_set: {
-                    shop_money: {
-                      amount: '729.95',
-                      currency_code: 'USD',
-                    },
-                    presentment_money: {
-                      amount: '729.95',
-                      currency_code: 'USD',
-                    },
-                  },
-                  line_price_set: {
-                    shop_money: {
-                      amount: '2189.85',
-                      currency_code: 'USD',
-                    },
-                    presentment_money: {
-                      amount: '2189.85',
-                      currency_code: 'USD',
-                    },
-                  },
-                  original_line_price_set: {
-                    shop_money: {
-                      amount: '2189.85',
-                      currency_code: 'USD',
-                    },
-                    presentment_money: {
-                      amount: '2189.85',
-                      currency_code: 'USD',
-                    },
-                  },
-                  price_set: {
-                    shop_money: {
-                      amount: '729.95',
-                      currency_code: 'USD',
-                    },
-                    presentment_money: {
-                      amount: '729.95',
-                      currency_code: 'USD',
-                    },
-                  },
-                  total_discount_set: {
-                    shop_money: {
-                      amount: '0.0',
-                      currency_code: 'USD',
-                    },
-                    presentment_money: {
-                      amount: '0.0',
-                      currency_code: 'USD',
-                    },
-                  },
-                },
-              ],
-              note: '',
-              updated_at: '2024-09-17T08:15:13.280Z',
-              created_at: '2024-09-16T03:50:15.478Z',
-              query_parameters: {
-                topic: ['carts_update'],
-                writeKey: ['2l9QoM7KRMJLMcYhXNUVDT0Mqbd'],
-              },
-            },
-            source: dummySourceConfig,
-          },
-        ],
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-      },
-      pathSuffix: '',
-    },
-    output: {
-      response: {
-        status: 200,
-        body: [
-          {
-            output: {
-              batch: [
-                {
-                  context: {
-                    library: {
-                      name: 'RudderStack Shopify Cloud',
-                      eventOrigin: 'server',
-                      version: '2.0.0',
-                    },
-                    integration: {
-                      name: 'SHOPIFY',
-                    },
-                    topic: 'carts_update',
-                  },
-                  integrations: {
-                    SHOPIFY: true,
-                  },
-                  type: 'track',
-                  event: 'Cart Update',
-                  properties: {
-                    id: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-                    token: 'Z2NwLXVzLWVhc3QxOjAxSjdXRjdOQjY0NlFFNFdQVEg0MTRFM1E2',
-                    note: '',
-                    updated_at: '2024-09-17T08:15:13.280Z',
-                    created_at: '2024-09-16T03:50:15.478Z',
-                    products: [
-                      {
-                        product_id: 7234590638193,
-                        price: '729.95',
-                        brand: 'pixel-testing-rs',
-                        quantity: 3,
-                        id: 41327143059569,
-                        properties: null,
-                        key: '41327143059569:90562f18109e0e6484b0c297e7981b30',
-                        discounted_price: '729.95',
-                        discounts: [],
-                        gift_card: false,
-                        grams: 0,
-                        line_price: '2189.85',
-                        original_line_price: '2189.85',
-                        original_price: '729.95',
-                        taxable: true,
-                        title: 'The Multi-location Snowboard',
-                        total_discount: '0.00',
-                        discounted_price_set: {
-                          shop_money: {
-                            amount: '729.95',
-                            currency_code: 'USD',
-                          },
-                          presentment_money: {
-                            amount: '729.95',
-                            currency_code: 'USD',
-                          },
-                        },
-                        line_price_set: {
-                          shop_money: {
-                            amount: '2189.85',
-                            currency_code: 'USD',
-                          },
-                          presentment_money: {
-                            amount: '2189.85',
-                            currency_code: 'USD',
-                          },
-                        },
-                        original_line_price_set: {
-                          shop_money: {
-                            amount: '2189.85',
-                            currency_code: 'USD',
-                          },
-                          presentment_money: {
-                            amount: '2189.85',
-                            currency_code: 'USD',
-                          },
-                        },
-                        price_set: {
-                          shop_money: {
-                            amount: '729.95',
-                            currency_code: 'USD',
-                          },
-                          presentment_money: {
-                            amount: '729.95',
-                            currency_code: 'USD',
-                          },
-                        },
-                        total_discount_set: {
-                          shop_money: {
-                            amount: '0.0',
-                            currency_code: 'USD',
-                          },
-                          presentment_money: {
-                            amount: '0.0',
-                            currency_code: 'USD',
-                          },
-                        },
-                        variant: '41327143059569  ',
-                      },
-                    ],
-                  },
-                  anonymousId: '97fcd7b2-cc24-47d7-b776-057b7b199513',
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-    mockFns: () => {
-      defaultMockFns();
-    },
-  },
-];
-
 export const data = [
+  ...pixelCheckoutEventsTestScenarios,
+  ...pixelCheckoutStepsScenarios,
   ...pixelEventsTestScenarios,
   ...serverSideEventsScenarios,
   ...v1ServerSideEventsScenarios,
