+++ test/integrations/sources/shopify/pixelTestScenarios/CheckoutEventsTests.ts
@@ -1,0 +1,570 @@
+// This file contains the test scenarios for the pixel checkout events
+import { dummySourceConfig, dummyBillingAddresses, dummyContext } from '../constants';
+
+export const pixelCheckoutEventsTestScenarios = [
+  {
+    name: 'shopify',
+    description: 'Track Call -> checkout_started event from web pixel',
+    module: 'source',
+    version: 'v1',
+    input: {
+      request: {
+        body: [
+          {
+            event: {
+              id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
+              name: 'checkout_started',
+              data: {
+                checkout: {
+                  buyerAcceptsEmailMarketing: false,
+                  buyerAcceptsSmsMarketing: false,
+                  attributes: [],
+                  billingAddress: dummyBillingAddresses[0],
+                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
+                  currencyCode: 'USD',
+                  discountApplications: [],
+                  discountsAmount: {
+                    amount: 0,
+                    currencyCode: 'USD',
+                  },
+                  email: '',
+                  phone: '',
+                  lineItems: [
+                    {
+                      discountAllocations: [],
+                      id: '41327143321713',
+                      quantity: 2,
+                      title: 'The Collection Snowboard: Liquid',
+                      variant: {
+                        id: '41327143321713',
+                        image: {
+                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
+                        },
+                        price: {
+                          amount: 749.95,
+                          currencyCode: 'USD',
+                        },
+                        product: {
+                          id: '7234590834801',
+                          title: 'The Collection Snowboard: Liquid',
+                          vendor: 'Hydrogen Vendor',
+                          type: 'snowboard',
+                          untranslatedTitle: 'The Collection Snowboard: Liquid',
+                          url: '/products/the-collection-snowboard-liquid',
+                        },
+                        sku: null,
+                        title: null,
+                        untranslatedTitle: null,
+                      },
+                      finalLinePrice: {
+                        amount: 1499.9,
+                        currencyCode: 'USD',
+                      },
+                      sellingPlanAllocation: null,
+                      properties: [],
+                    },
+                  ],
+                  localization: {
+                    country: {
+                      isoCode: 'US',
+                    },
+                    language: {
+                      isoCode: 'en-US',
+                    },
+                    market: {
+                      id: 'gid://shopify/Market/23505895537',
+                      handle: 'us',
+                    },
+                  },
+                  order: {
+                    id: null,
+                    customer: {
+                      id: null,
+                      isFirstOrder: null,
+                    },
+                  },
+                  delivery: {
+                    selectedDeliveryOptions: [],
+                  },
+                  shippingAddress: dummyBillingAddresses[0],
+                  subtotalPrice: {
+                    amount: 2759.8,
+                    currencyCode: 'USD',
+                  },
+                  shippingLine: {
+                    price: {
+                      amount: 0,
+                      currencyCode: 'USD',
+                    },
+                  },
+                  smsMarketingPhone: null,
+                  totalTax: {
+                    amount: 0,
+                    currencyCode: 'USD',
+                  },
+                  totalPrice: {
+                    amount: 2759.8,
+                    currencyCode: 'USD',
+                  },
+                  transactions: [],
+                },
+              },
+              type: 'standard',
+              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
+              timestamp: '2024-09-15T20:57:59.674Z',
+              context: dummyContext,
+              pixelEventLabel: true,
+              query_parameters: {
+                topic: ['checkouts_update'],
+                writeKey: ['dummy-write-key'],
+              },
+            },
+            source: dummySourceConfig,
+          },
+        ],
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+      },
+      pathSuffix: '',
+    },
+    output: {
+      response: {
+        status: 200,
+        body: [
+          {
+            output: {
+              batch: [
+                {
+                  context: {
+                    document: {
+                      location: {
+                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        hash: '',
+                        host: 'store.myshopify.com',
+                        hostname: 'store.myshopify.com',
+                        origin: 'https://store.myshopify.com',
+                        pathname:
+                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        port: '',
+                        protocol: 'https:',
+                        search: '',
+                      },
+                      referrer: 'https://store.myshopify.com/cart',
+                      characterSet: 'UTF-8',
+                      title: 'Checkout - pixel-testing-rs',
+                    },
+                    navigator: {
+                      language: 'en-US',
+                      cookieEnabled: true,
+                      languages: ['en-US', 'en'],
+                      userAgent:
+                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
+                    },
+                    window: {
+                      innerHeight: 1028,
+                      innerWidth: 1362,
+                      outerHeight: 1080,
+                      outerWidth: 1728,
+                      pageXOffset: 0,
+                      pageYOffset: 0,
+                      location: {
+                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        hash: '',
+                        host: 'store.myshopify.com',
+                        hostname: 'store.myshopify.com',
+                        origin: 'https://store.myshopify.com',
+                        pathname:
+                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        port: '',
+                        protocol: 'https:',
+                        search: '',
+                      },
+                      origin: 'https://store.myshopify.com',
+                      screen: {
+                        height: 1117,
+                        width: 1728,
+                      },
+                      screenX: 0,
+                      screenY: 37,
+                      scrollX: 0,
+                      scrollY: 0,
+                    },
+                    page: {
+                      title: 'Checkout - pixel-testing-rs',
+                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                      search: '',
+                    },
+                    userAgent:
+                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
+                    screen: {
+                      height: 1117,
+                      width: 1728,
+                    },
+                    library: {
+                      name: 'RudderStack Shopify Cloud',
+                      eventOrigin: 'client',
+                      version: '2.0.0',
+                    },
+                    topic: 'checkout_started',
+                  },
+                  integrations: {
+                    SHOPIFY: true,
+                  },
+                  type: 'track',
+                  event: 'Checkout Started',
+                  properties: {
+                    products: [
+                      {
+                        discountAllocations: [],
+                        id: '41327143321713',
+                        quantity: 2,
+                        title: 'The Collection Snowboard: Liquid',
+                        variant: 'The Collection Snowboard: Liquid',
+                        finalLinePrice: {
+                          amount: 1499.9,
+                          currencyCode: 'USD',
+                        },
+                        sellingPlanAllocation: null,
+                        properties: [],
+                        name: 'The Collection Snowboard: Liquid',
+                        image_url:
+                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
+                        price: 749.95,
+                        sku: null,
+                        product_id: '7234590834801',
+                        category: 'snowboard',
+                        url: '/products/the-collection-snowboard-liquid',
+                        brand: 'Hydrogen Vendor',
+                      },
+                    ],
+                    order_id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
+                    checkout_id: '5f7028e0bd5225c17b24bdaa0c09f914',
+                    total: 2759.8,
+                    currency: 'USD',
+                    discount: 0,
+                    shipping: 0,
+                    revenue: 2759.8,
+                    value: 2759.8,
+                    tax: 0,
+                  },
+                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
+                },
+              ],
+            },
+          },
+        ],
+      },
+    },
+  },
+  {
+    name: 'shopify',
+    description: 'Track Call -> checkout_completed event from web pixel',
+    module: 'source',
+    version: 'v1',
+    input: {
+      request: {
+        body: [
+          {
+            event: {
+              id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
+              name: 'checkout_completed',
+              data: {
+                checkout: {
+                  buyerAcceptsEmailMarketing: false,
+                  buyerAcceptsSmsMarketing: false,
+                  attributes: [],
+                  billingAddress: dummyBillingAddresses[0],
+                  token: '5f7028e0bd5225c17b24bdaa0c09f914',
+                  currencyCode: 'USD',
+                  discountApplications: [],
+                  discountsAmount: {
+                    amount: 0,
+                    currencyCode: 'USD',
+                  },
+                  email: '',
+                  phone: '',
+                  lineItems: [
+                    {
+                      discountAllocations: [],
+                      id: '41327143321713',
+                      quantity: 2,
+                      title: 'The Collection Snowboard: Liquid',
+                      variant: {
+                        id: '41327143321713',
+                        image: {
+                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
+                        },
+                        price: {
+                          amount: 749.95,
+                          currencyCode: 'USD',
+                        },
+                        product: {
+                          id: '7234590834801',
+                          title: 'The Collection Snowboard: Liquid',
+                          vendor: 'Hydrogen Vendor',
+                          type: 'snowboard',
+                          untranslatedTitle: 'The Collection Snowboard: Liquid',
+                          url: '/products/the-collection-snowboard-liquid',
+                        },
+                        sku: null,
+                        title: null,
+                        untranslatedTitle: null,
+                      },
+                      finalLinePrice: {
+                        amount: 1499.9,
+                        currencyCode: 'USD',
+                      },
+                      sellingPlanAllocation: null,
+                      properties: [],
+                    },
+                    {
+                      discountAllocations: [],
+                      id: '41327143157873',
+                      quantity: 2,
+                      title: 'The Multi-managed Snowboard',
+                      variant: {
+                        id: '41327143157873',
+                        image: {
+                          src: 'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
+                        },
+                        price: {
+                          amount: 629.95,
+                          currencyCode: 'USD',
+                        },
+                        product: {
+                          id: '7234590736497',
+                          title: 'The Multi-managed Snowboard',
+                          vendor: 'Multi-managed Vendor',
+                          type: 'snowboard',
+                          untranslatedTitle: 'The Multi-managed Snowboard',
+                          url: '/products/the-multi-managed-snowboard',
+                        },
+                        sku: 'sku-managed-1',
+                        title: null,
+                        untranslatedTitle: null,
+                      },
+                      finalLinePrice: {
+                        amount: 1259.9,
+                        currencyCode: 'USD',
+                      },
+                      sellingPlanAllocation: null,
+                      properties: [],
+                    },
+                  ],
+                  localization: {
+                    country: {
+                      isoCode: 'US',
+                    },
+                    language: {
+                      isoCode: 'en-US',
+                    },
+                    market: {
+                      id: 'gid://shopify/Market/23505895537',
+                      handle: 'us',
+                    },
+                  },
+                  order: {
+                    id: null,
+                    customer: {
+                      id: null,
+                      isFirstOrder: null,
+                    },
+                  },
+                  delivery: {
+                    selectedDeliveryOptions: [],
+                  },
+                  shippingAddress: dummyBillingAddresses[0],
+                  subtotalPrice: {
+                    amount: 2759.8,
+                    currencyCode: 'USD',
+                  },
+                  shippingLine: {
+                    price: {
+                      amount: 0,
+                      currencyCode: 'USD',
+                    },
+                  },
+                  smsMarketingPhone: null,
+                  totalTax: {
+                    amount: 0,
+                    currencyCode: 'USD',
+                  },
+                  totalPrice: {
+                    amount: 2759.8,
+                    currencyCode: 'USD',
+                  },
+                  transactions: [],
+                },
+              },
+              type: 'standard',
+              clientId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
+              timestamp: '2024-09-15T20:57:59.674Z',
+              context: dummyContext,
+              pixelEventLabel: true,
+              query_parameters: {
+                topic: ['checkouts_update'],
+                writeKey: ['dummy-write-key'],
+              },
+            },
+            source: dummySourceConfig,
+          },
+        ],
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+      },
+      pathSuffix: '',
+    },
+    output: {
+      response: {
+        status: 200,
+        body: [
+          {
+            output: {
+              batch: [
+                {
+                  context: {
+                    document: {
+                      location: {
+                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        hash: '',
+                        host: 'store.myshopify.com',
+                        hostname: 'store.myshopify.com',
+                        origin: 'https://store.myshopify.com',
+                        pathname:
+                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        port: '',
+                        protocol: 'https:',
+                        search: '',
+                      },
+                      referrer: 'https://store.myshopify.com/cart',
+                      characterSet: 'UTF-8',
+                      title: 'Checkout - pixel-testing-rs',
+                    },
+                    navigator: {
+                      language: 'en-US',
+                      cookieEnabled: true,
+                      languages: ['en-US', 'en'],
+                      userAgent:
+                        'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
+                    },
+                    window: {
+                      innerHeight: 1028,
+                      innerWidth: 1362,
+                      outerHeight: 1080,
+                      outerWidth: 1728,
+                      pageXOffset: 0,
+                      pageYOffset: 0,
+                      location: {
+                        href: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        hash: '',
+                        host: 'store.myshopify.com',
+                        hostname: 'store.myshopify.com',
+                        origin: 'https://store.myshopify.com',
+                        pathname:
+                          '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                        port: '',
+                        protocol: 'https:',
+                        search: '',
+                      },
+                      origin: 'https://store.myshopify.com',
+                      screen: {
+                        height: 1117,
+                        width: 1728,
+                      },
+                      screenX: 0,
+                      screenY: 37,
+                      scrollX: 0,
+                      scrollY: 0,
+                    },
+                    page: {
+                      title: 'Checkout - pixel-testing-rs',
+                      url: 'https://store.myshopify.com/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                      path: '/checkouts/cn/Z2NwLXVzLWVhc3QxOjAxSjY5OVpIRURQNERFMDBKUTVaRkI4UzdU',
+                      search: '',
+                    },
+                    userAgent:
+                      'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36',
+                    screen: {
+                      height: 1117,
+                      width: 1728,
+                    },
+                    library: {
+                      name: 'RudderStack Shopify Cloud',
+                      eventOrigin: 'client',
+                      version: '2.0.0',
+                    },
+                    topic: 'checkout_completed',
+                  },
+                  integrations: {
+                    SHOPIFY: true,
+                  },
+                  type: 'track',
+                  event: 'Checkout Completed',
+                  properties: {
+                    products: [
+                      {
+                        discountAllocations: [],
+                        id: '41327143321713',
+                        quantity: 2,
+                        title: 'The Collection Snowboard: Liquid',
+                        variant: 'The Collection Snowboard: Liquid',
+                        finalLinePrice: {
+                          amount: 1499.9,
+                          currencyCode: 'USD',
+                        },
+                        sellingPlanAllocation: null,
+                        properties: [],
+                        name: 'The Collection Snowboard: Liquid',
+                        image_url:
+                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_b13ad453-477c-4ed1-9b43-81f3345adfd6_64x64.jpg?v=1724736600',
+                        price: 749.95,
+                        sku: null,
+                        product_id: '7234590834801',
+                        category: 'snowboard',
+                        url: '/products/the-collection-snowboard-liquid',
+                        brand: 'Hydrogen Vendor',
+                      },
+                      {
+                        discountAllocations: [],
+                        id: '41327143157873',
+                        quantity: 2,
+                        title: 'The Multi-managed Snowboard',
+                        variant: 'The Multi-managed Snowboard',
+                        finalLinePrice: {
+                          amount: 1259.9,
+                          currencyCode: 'USD',
+                        },
+                        sellingPlanAllocation: null,
+                        properties: [],
+                        name: 'The Multi-managed Snowboard',
+                        image_url:
+                          'https://cdn.shopify.com/s/files/1/0590/2696/4593/files/Main_9129b69a-0c7b-4f66-b6cf-c4222f18028a_64x64.jpg?v=1724736597',
+                        price: 629.95,
+                        sku: 'sku-managed-1',
+                        product_id: '7234590736497',
+                        category: 'snowboard',
+                        url: '/products/the-multi-managed-snowboard',
+                        brand: 'Multi-managed Vendor',
+                      },
+                    ],
+                    order_id: 'sh-f77a78f1-C1D8-4ED4-9C9B-0D352CF6F3BF',
+                    checkout_id: '5f7028e0bd5225c17b24bdaa0c09f914',
+                    total: 2759.8,
+                    currency: 'USD',
+                    discount: 0,
+                    shipping: 0,
+                    revenue: 2759.8,
+                    value: 2759.8,
+                    tax: 0,
+                  },
+                  anonymousId: 'c7b3f99b-4d34-463b-835f-c879482a7750',
+                },
+              ],
+            },
+          },
+        ],
+      },
+    },
+  },
+];
